1:新增demuxer中部分seek相关逻辑 2:涉及到的flush操作，demuxer相关逻辑中已经对解码器加锁，   但音、视频线程中暂时没加，接下来首要任务就是加锁。

+++ src/tools.cpp
@@ -92,6 +92,6 @@ void swsDeleter(SwsContext* swsPtr){
 }
 
 double getCurPts(SyncClock& clock){
-    double curTime = av_gettime_relative()/1e6;
+    double curTime = (double)av_gettime_relative()/1e6;
     return curTime + clock.diff;
 }
